requirements.txt
easyocr==1.6.2
numpy==1.22.3
opencv_python_headless==4.4.0.46
pandas==1.4.3
Pillow==9.3.0
streamlit==1.24.1

